# MRO/openapi/mms.yaml
openapi: 3.0.3
info:
  title: Maintenance Management System (MMS) API
  version: 1.0.0
  description: API for work orders execution, tasks check-lists, preventive maintenance schedules, and maintenance history.
servers:
  - url: /api/v1
paths:
...
  /mms/work-orders:
    get:
      summary: List work orders
      operationId: getWorkOrders
      security:
        - basicAuth: []
      responses:
        '200':
          description: A list of work orders.
          content:
            application/json:
              schema:
                type: object
                properties:
                  meta:
                    $ref: '#/components/schemas/ApiMeta'
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/WorkOrderResponse'
        '401':
          $ref: '#/components/responses/401Unauthorized'
        '403':
          $ref: '#/components/responses/403Forbidden'
    post:
      summary: Create work order
      operationId: createWorkOrder
      security:
        - basicAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateWorkOrderRequest'
      responses:
        '201':
          description: Work order created.
          content:
            application/json:
              schema:
                type: object
                properties:
                  meta:
                    $ref: '#/components/schemas/ApiMeta'
                  data:
                    $ref: '#/components/schemas/WorkOrderResponse'
        '400':
          $ref: '#/components/responses/400BadRequest'
        '401':
          $ref: '#/components/responses/401Unauthorized'
        '403':
          $ref: '#/components/responses/403Forbidden'
...
components:
  securitySchemes:
    basicAuth:
      type: http
      scheme: basic
  schemas:
    ApiMeta:
      type: object
      properties:
        requestId:
          type: string
          nullable: true
    ApiErrorResponse:
      type: object
      properties:
        timestamp:
          type: string
          format: date-time
        status:
          type: integer
        error:
          type: string
        message:
          type: string
        path:
          type: string
        requestId:
          type: string
          nullable: true
...
    CreateWorkOrderRequest:
      type: object
      required:
        - woNumber
        - equipmentId
        - type
        - title
      properties:
        woNumber:
          type: string
          maxLength: 64
        equipmentId:
          type: string
          format: uuid
        type:
          type: string
          maxLength: 32
        priority:
          type: string
          maxLength: 32
          nullable: true
        scheduledDate:
          type: string
          format: date-time
          nullable: true
        title:
          type: string
          maxLength: 255
        description:
          type: string
          maxLength: 2000
          nullable: true
...
    WorkOrderResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
        woNumber:
          type: string
        equipmentId:
          type: string
          format: uuid
        type:
          type: string
        priority:
          type: string
        status:
          type: string
          enum: [OPEN, PLANNED, ASSIGNED, IN_PROGRESS, WAITING_PARTS, COMPLETED, CANCELLED]
        scheduledDate:
          type: string
          format: date-time
          nullable: true
        startedAt:
          type: string
          format: date-time
          nullable: true
        completedDate:
          type: string
          format: date-time
          nullable: true
        technicianId:
          type: string
          format: uuid
          nullable: true
        title:
          type: string
        description:
          type: string
          nullable: true
        completionAct:
          type: string
          nullable: true
        signatureHash:
          type: string
          nullable: true
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
  responses:
    400BadRequest:
      description: Invalid request parameters or input payload.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiErrorResponse'
    401Unauthorized:
      description: Basic authentication credentials missing or invalid.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiErrorResponse'
    403Forbidden:
      description: Authenticated user does not have the required permissions.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiErrorResponse'
...